+++ lib/src/shared/status.dart
@@ -18,7 +18,6 @@
 import 'dart:convert';
 
 import 'package:grpc/src/generated/google/protobuf/any.pb.dart';
-import 'package:grpc/src/generated/google/rpc/code.pb.dart';
 import 'package:grpc/src/generated/google/rpc/error_details.pb.dart';
 import 'package:grpc/src/generated/google/rpc/status.pb.dart';
 import 'package:meta/meta.dart';
@@ -151,6 +150,28 @@ class StatusCode {
   static int fromHttpStatus(int status) {
     return _httpStatusToGrpcStatus[status] ?? StatusCode.unknown;
   }
+
+  /// Creates a string from a gRPC status code.
+  static String? name(int status) => switch (status) {
+        ok => 'OK',
+        cancelled => 'CANCELLED',
+        unknown => 'UNKNOWN',
+        invalidArgument => 'INVALID_ARGUMENT',
+        deadlineExceeded => 'DEADLINE_EXCEEDED',
+        notFound => 'NOT_FOUND',
+        alreadyExists => 'ALREADY_EXISTS',
+        permissionDenied => 'PERMISSION_DENIED',
+        resourceExhausted => 'RESOURCE_EXHAUSTED',
+        failedPrecondition => 'FAILED_PRECONDITION',
+        aborted => 'ABORTED',
+        outOfRange => 'OUT_OF_RANGE',
+        unimplemented => 'UNIMPLEMENTED',
+        internal => 'INTERNAL',
+        unavailable => 'UNAVAILABLE',
+        dataLoss => 'DATA_LOSS',
+        unauthenticated => 'UNAUTHENTICATED',
+        int() => null,
+      };
 }
 
 class GrpcError implements Exception {
@@ -308,7 +329,8 @@ class GrpcError implements Exception {
         code = StatusCode.unauthenticated;
 
   /// Given a status code, return the name
-  String get codeName => (Code.valueOf(code) ?? Code.UNKNOWN).name;
+  String get codeName =>
+      StatusCode.name(code) ?? StatusCode.name(StatusCode.unknown)!;
 
   @override
   bool operator ==(other) {